Fixing form object creation error handling
Previously, the try/catch block around the form object creation function, called on form initialization, was not firing properly. Upon investigation it looks like the try/catch was not the proper way of catching the error, as it is an async function. Using .catch() instead worked to properly handle errors coming from the form library.