Fixing issue with missing classes for some components
Multiple and Group components were previously rendering incorrectly. This is part of a refactor where some class names changed, but are now tied more directly to Multiple and Group components instead of relying on child elements. This also allows a more consistent look and feel.

Classes were added for group and multiple fields, and added to both the edit and view components