Fixed WebID login bug
The issue was that the login page made the assumption that the WebID was hosted on the same origin as the Identity Provider, which is not necessarily the case. This fix adds a check on the `solid:oidcIssuer` predicate in the Profile Document to get the Identity Provider address, while remaining backwards compatible by also testing the presence of the IdP at the WebID origin.